+++ code-pushup.preset.ts
@@ -1,7 +1,10 @@
 /* eslint-disable @nx/enforce-module-boundaries */
 import { createProjectGraphAsync } from '@nx/devkit';
 import type { CoreConfig, PluginUrls } from './packages/models/src/index.js';
-import axePlugin, { axeGroupRefs } from './packages/plugin-axe/src/index.js';
+import axePlugin, {
+  type AxePluginOptions,
+  axeGroupRefs,
+} from './packages/plugin-axe/src/index.js';
 import coveragePlugin, {
   type CoveragePluginConfig,
   getNxCoveragePaths,
@@ -217,8 +220,11 @@ export async function configureLighthousePlugin(
   };
 }
 
-export function configureAxePlugin(urls: PluginUrls): CoreConfig {
-  const axe = axePlugin(urls);
+export function configureAxePlugin(
+  urls: PluginUrls,
+  options?: AxePluginOptions,
+): CoreConfig {
+  const axe = axePlugin(urls, options);
   return {
     plugins: [axe],
     categories: [